PitiviRuler, PitiviTimelineWindow : better refresh for ruler
git-svn-id: svn+ssh://svn.gnome.org/svn/pitivi/trunk@579 d3729300-e425-0410-8a4c-d956edccc248

+++ src/pitivi-ruler.h
@@ -100,14 +100,14 @@ typedef struct _PitiviRulerPrivate PitiviRulerPrivate;
 struct _PitiviRuler
 {
   GtkRuler  ruler;
-  guint	    time_pix;
+  gint64	    time_pix;
   PitiviRulerPrivate *private;
 };
 
 struct _PitiviRulerClass
 {
   GtkRulerClass parent_class;
-  void (* moving) (PitiviRuler *ruler, gint *gstep);
+  void (* moving) (PitiviRuler *ruler, gint64 *gstep);
 };
